webapps2024/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==4.2.11
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-crispy-forms==2.1
djangorestframework==3.15.0
idna==3.7
PyJWT==1.7.1
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
